+++ src/repositories/base/baseMongoPromiseRepository.ts
@@ -14,6 +14,17 @@ export class BaseMongoPromiseRepository implements MongoPromiseRepositoryInterfa
         return this;
     }
 
+    openSingletonConnection(): Promise<any> {
+        return new Promise((resolve, reject) => {
+            this.mongoRep.openSingletonConnection((err: any) => {
+                if (err)
+                    reject(err);
+                else
+                    resolve("Connection Opened");
+            })
+        });
+    }
+
     closeSingletonConnection(): Promise<any> {
         return new Promise((resolve, reject) => {
             this.mongoRep.closeSingletonConnection((err: any) => {